feat:(product.proudctInfo,productInfoEs):商品销量添加异步Redis缓冲，elasticsearch修改商品信息

+++ neko-transaction-product/src/main/java/neko/transaction/product/mapper/ProductInfoMapper.java
@@ -78,5 +78,5 @@ public interface ProductInfoMapper extends BaseMapper<ProductInfo> {
      * @param productId 商品id
      * @param increase 要添加的数量
      */
-    void increaseSaleNumber(String productId, Integer increase);
+    void increaseSaleNumber(String productId, Long increase);
 }
